+++ src/instructlab/log.py
@@ -11,9 +11,9 @@ FORMAT = "%(levelname)s %(asctime)s %(filename)s:%(lineno)d: %(funcName)s %(mess
 class CustomFormatter(logging.Formatter):
     def format(self, record):
         if record.filename == os.path.basename(__file__):
-            # Replace our current file name with 'llama_cpp' since the message is coming from stdout or
+            # Replace our current file name with 'serving_backend' since the message is coming from stdout or
             # stderr. Printing this file name is not useful in this case, we need to print where the line is coming from.
-            record.filename = "llama_cpp"
+            record.filename = "serving_backend"
             record.lineno = 0  # Remove line number, no point of showing it
         return super().format(record)